pyproject.toml
[tool.poetry]
name = "pretty-coursework"
version = "0.1.0"
description = "Buranya"
authors = ["Bogdanchik"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
xlrd = "^2.0.1"


[tool.poetry.group.lint.dependencies]
mypy = "^1.10.1"
flake8 = "^7.1.0"
isort = "^5.13.2"
black = "^24.4.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py
)
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
